backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
python-jose==3.3.0
bcrypt==4.1.2
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
marshmallow==3.20.2
google-genai 